move script files to own folder, improve error catching in `sh` scripts

+++ scripts/upgrade-install-plugin.sh
@@ -33,35 +33,34 @@ if ! command -v jq > /dev/null; then
   echo "Failed to find required jq command."
 fi
 
-tmp_dir=$(mktemp -d -t $TARGET_PLUGIN_BUNDLE_NAME.XXXXXXX)
-if [ "$?" != "0" ]; then
+tmp_dir=$(mktemp -d -t "$TARGET_PLUGIN_BUNDLE_NAME".XXXXXXX)
+if ! [ "$tmp_dir" ]; then
   echo "Failed to create temporary directory."
   exit 1
 fi
 
 echo "Downloading ${TARGET_PLUGIN_BUNDLE_NAME}.sha256..."
-curl -fsSL# -o "$tmp_dir/SHASUMS256.txt" \
-  "$REPO_BASE_URL/${TARGET_PLUGIN_BUNDLE_NAME}.sha256"
-if [ "$?" != "0" ]; then
+if ! curl -fsSL# -o "$tmp_dir/SHASUMS256.txt" "$REPO_BASE_URL/${TARGET_PLUGIN_BUNDLE_NAME}.sha256"; then
   echo "Failed to download SHASUMS256.txt"
   exit 1
 fi
 
 echo "Downloading ${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz..."
-curl -fL# -o "$tmp_dir/${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz" \
-  "$REPO_BASE_URL/${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz"
-if [ "$?" != "0" ]; then
+if ! curl -fL# -o "$tmp_dir/${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz" "$REPO_BASE_URL/${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz"; then
   echo "Failed to download homebridge-config-ui-x-${TARGET_VERSION}.tar.gz"
   exit 1
 fi
 echo ""
 
 echo "Verifying download..."
-cd $tmp_dir
+if ! cd "$tmp_dir"; then
+  echo "Failed to change directory to $tmp_dir"
+  exit 1
+fi
 $SHASUM_COMMAND -c SHASUMS256.txt
-if [ "$?" != "0" ]; then
+if ! $SHASUM_COMMAND -c SHASUMS256.txt; then
   echo "Download failed integrity check."
-  rm -rf $tmp_dir
+  rm -rf "$tmp_dir"
   exit 1
 fi
 echo ""
@@ -77,9 +76,9 @@ if [ -d "$TARGET_PATH/$TARGET_PLUGIN" ]; then
 
   if [ -d "$TARGET_PATH/.bin/" ]; then
     for FILE in "$TARGET_PATH/.bin/"*; do
-      if [[ "$(readlink $FILE)" == "../$TARGET_PLUGIN/"* ]]; then
+      if [[ "$(readlink "$FILE")" == "../$TARGET_PLUGIN/"* ]]; then
         mkdir -p "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bin.bak"
-        mv $FILE "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bin.bak/"
+        mv "$FILE" "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bin.bak/"
       fi
     done
   fi
@@ -89,11 +88,11 @@ echo ""
 function postInstallCleanup() {
   rm -rf "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bak"
   rm -rf "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bin.bak/"
-  rm -rf $tmp_dir
+  rm -rf "$tmp_dir"
 }
 
 function revertToBackup() {
-  rm -rf "$TARGET_PATH/$TARGET_PLUGIN"
+  rm -rf "${TARGET_PATH:?}/${TARGET_PLUGIN:?}"
   if [ -d "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bak" ]; then
     echo "Restoring previous version..."
     mv "$TARGET_PATH/.$TARGET_PLUGIN_BUNDLE_NAME.bak" "$TARGET_PATH/$TARGET_PLUGIN"
@@ -105,8 +104,7 @@ function revertToBackup() {
 }
 
 echo "Extracting..."
-tar --no-same-owner -xmvf "$tmp_dir/${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz" -C "$TARGET_PATH"
-if [ "$?" != "0" ]; then
+if ! tar --no-same-owner -xmvf "$tmp_dir/${TARGET_PLUGIN_BUNDLE_NAME}.tar.gz" -C "$TARGET_PATH"; then
   echo "Failed to extract."
   revertToBackup
 else
@@ -115,23 +113,27 @@ fi
 echo ""
 
 echo "Running post-install scripts..."
-cd "$TARGET_PATH/$TARGET_PLUGIN"
-npm rebuild --foreground-scripts
-if [ "$?" != "0" ]; then
+if ! cd "$TARGET_PATH/$TARGET_PLUGIN"; then
+  echo "Failed to change directory to $TARGET_PATH/$TARGET_PLUGIN"
+  exit 1
+fi
+if ! npm rebuild --foreground-scripts; then
   echo "Failed to rebuild."
   revertToBackup
 fi
 echo ""
 
-PLUGIN_PACKAGE_JSON_PATH="$(dirname $TARGET_PATH)/package.json"
+PLUGIN_PACKAGE_JSON_PATH="$(dirname "$TARGET_PATH")/package.json"
 
 if [ ! -f "$PLUGIN_PACKAGE_JSON_PATH" ]; then
   echo "{}" > "$PLUGIN_PACKAGE_JSON_PATH"
 fi
 echo "Updating $PLUGIN_PACKAGE_JSON_PATH..."
-PACKAGE_JSON=$(cat "$PLUGIN_PACKAGE_JSON_PATH" | jq -rM ".dependencies += { \"$TARGET_PLUGIN\": \"$TARGET_VERSION\"}")
-if [ "$?" = "0" ]; then
-  printf "$PACKAGE_JSON" > "$PLUGIN_PACKAGE_JSON_PATH"
+if PACKAGE_JSON=$(jq -rM ".dependencies += { \"$TARGET_PLUGIN\": \"$TARGET_VERSION\"}" < "$PLUGIN_PACKAGE_JSON_PATH"); then
+  printf "%s" "$PACKAGE_JSON" > "$PLUGIN_PACKAGE_JSON_PATH"
+else
+  echo "Failed to update $PLUGIN_PACKAGE_JSON_PATH"
+  exit 1
 fi
 echo ""
